+++ lib/src/requests/responses/restrr_errors.dart
@@ -2,8 +2,17 @@ import 'package:restrr/restrr.dart';
 
 /// Represents a type of error that can occur during a REST request.
 enum RestrrError {
+  /* Request errors */
+  badRequest,
+
+  /* Server errors */
+
+  notFound,
+  notSignedIn,
   invalidCredentials,
   alreadySignedIn,
+  internalServerError,
+  serviceUnavailable,
 
   /* Client errors */
 
@@ -16,16 +25,5 @@ enum RestrrError {
 
   const RestrrError({this.clientError = false});
 
-  RestResponse<T> toRestResponse<T>({int? statusCode}) {
-    return ErrorResponse(type: this, statusCode: statusCode).toRestResponse(statusCode: statusCode);
-  }
-}
-
-class ErrorResponse {
-  final RestrrError type;
-  final int? statusCode;
-
-  const ErrorResponse({required this.type, this.statusCode});
-
-  RestResponse<T> toRestResponse<T>({int? statusCode}) => RestResponse(error: this);
+  RestResponse<T> toRestResponse<T>() => RestResponse(error: this);
 }