+++ lib/src/requests/route.dart
@@ -32,18 +32,6 @@ class Route {
 
   Route.patch(String path, {bool isVersioned = true}) : this._('PATCH', path, isVersioned: isVersioned);
 
-  /// Translates a [DioException] into a [RestrrError].
-  static Future<RestResponse<T>> translateDioException<T>(DioException error) async {
-    RestrrError err = RestrrError.unknown;
-    if (!await IOUtils.checkConnection()) {
-      err = RestrrError.noInternetConnection;
-    }
-    if (error.type == DioExceptionType.connectionTimeout || error.type == DioExceptionType.receiveTimeout) {
-      err = RestrrError.serverUnreachable;
-    }
-    return err.toRestResponse(statusCode: error.response?.statusCode);
-  }
-
   CompiledRoute compile({List<String> params = const []}) {
     if (params.length != paramCount) {
       throw ArgumentError(